added tests for switching to sidechain
